+++ react/components/ProductImages/constants/productImagesClasses.js
@@ -3,13 +3,10 @@ export default {
   VERTICAL_COMPONENT: 'vtex-product-image__vertical',
   HORIZONTAL_COMPONENT: 'vtex-product-image__horizontal',
   SELECTED_IMAGE: 'vtex-product-image__selected-image',
-  VERTICAL_THUMBNAIL_SLIDER: 'vtex-product-image__vertical-thumbnail-slider',
-  HORIZONTAL_THUMBNAIL_SLIDER:
-    'vtex-product-image__horizontal-thumbnail-slider',
+  THUMBNAIL_SLIDER: 'vtex-product-image__thumbnail-slider',
   THUMBNAIL_ITEM: 'vtex-product-image__thumbnail-slider-item',
-  VERTICAL_ARROW_PREV: 'vtex-product-image__vertical-arrow--prev',
-  VERTICAL_ARROW_NEXT: 'vtex-product-image__vertical-arrow--next',
-  HORIZONTAL_ARROW_PREV: 'vtex-product-image__horizontal-arrow--prev',
-  HORIZONTAL_ARROW_NEXT: 'vtex-product-image__horizontal-arrow--next',
+  IMAGE_ZOOM: 'vtex-product-image__zoom',
+  IMAGE_ZOOM_IMG: 'vtex-product-image__zoom-image',
+  IMAGE_ZOOM_CONTAINER: 'vtex-product-image__image-zoom-container',
   DOTS_CLASS: 'vtex-product-image__dots',
 }